01_etl/app/requirements.txt
django-cors-headers==3.13.0
django-debug-toolbar==3.5.0
django-extensions==3.2.0
django-split-settings==1.1.0
django==4.0.4
flake8==4.0
isort==5.10.1
more-itertools==8.13.0
psycopg2-binary==2.9
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.28.1
six==1.16.0
uWSGI==2.0.20